MODKBEKBJ-497 Fix raml errors
