+++ ramls/types/uc/costperuse/costPerUseParameters.json
@@ -8,11 +8,13 @@
   "properties": {
     "startMonth" : {
       "type": "string",
+      "description": "Fiscal starting month",
       "$ref": "../month.json",
       "example": "jan"
     },
     "currency": {
       "type": "string",
+      "description": "Currency",
       "example": "USD"
     }
   },
