+++ src/main/java/org/folio/rest/util/ExceptionMappers.java
@@ -21,7 +21,7 @@ import javax.ws.rs.NotAuthorizedException;
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.Response;
 
-import org.folio.client.uc.UCRequestException;
+import org.folio.client.uc.UCFailedRequestException;
 import org.folio.db.exc.AuthorizationException;
 import org.folio.db.exc.ConstraintViolationException;
 import org.folio.db.exc.DatabaseException;
@@ -110,7 +110,7 @@ public final class ExceptionMappers {
   }
 
   /**
-   * {@link UCRequestException} to {@link Response} error mapper
+   * {@link UCFailedRequestException} to {@link Response} error mapper
    * <pre>
    * Response.status = {@code 400}
    * Response.entity =  {@link org.folio.rest.jaxrs.model.JsonapiError}
@@ -119,7 +119,7 @@ public final class ExceptionMappers {
    *
    * @return mapper
    */
-  public static Function<UCRequestException, Response> error400UCRequestMapper() {
+  public static Function<UCFailedRequestException, Response> error400UCRequestMapper() {
     return exception ->
       Response.status(SC_BAD_REQUEST)
         .header(CONTENT_TYPE, JSON_API_TYPE)
